requirements.txt
langchain==0.3.26
langchain-community==0.3.27
chromadb==1.0.15
unstructured==0.18.9
pypdf==5.8.0
langchain-groq==0.3.6
gradio
sqlalchemy
tensorflow==2.18.0
tf-keras==2.18.0
tf_keras==2.18.0
transformers==4.53.2
huggingface-hub==0.33.2
sentence-transformers==4.1.0
keras==3.8.0
torch==2.6.0
absl-py==1.4.0
python-dotenv
psycopg2-binary